Consolidate leave calculation into single file with dynamic AJAX handling
